requirements.txt
alembic==1.17.2
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.1
cryptography==46.0.3
Flask==3.1.2
Flask-SocketIO==5.3.6
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
ifaddr==0.2.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
netaddr==1.3.0
packaging==25.0
pluggy==1.6.0
ply==3.11
pyasn1==0.6.0
pycparser==2.23
Pygments==2.19.2
pysmi-lextudio==1.4.3
pysnmp==7.1.22
pysnmpcrypto==0.0.4
pytest==9.0.2
pytest-cov==4.1.0
python-dotenv==1.2.1
python-engineio==4.11.1
python-socketio==5.11.1
requests==2.32.5
simple-websocket==1.1.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.4
wsproto==1.3.2
zeroconf==0.148.0
zope.event==6.1
zope.interface==8.1.1
